refactor: add computer repository save method
