+++ Program.cs
@@ -1,5 +1,6 @@
 ﻿using LabManager.Repositories;
 using LabManager.Database;
+using LabManager.Models;
 using Microsoft.Data.Sqlite;
 
 var modelName = args[0];
@@ -23,18 +24,9 @@ if(modelName == "Computer")
         var id = Convert.ToInt32(args[2]);
         var ram = args[3];
         var processor = args[4];
-        var connection = new SqliteConnection("Data Source=database.db");
-        connection.Open();
 
-        var command = connection.CreateCommand();
-        command.CommandText = "INSERT INTO Computers VALUES ($id, $ram, $processor);"
-        ;
-        command.Parameters.AddWithValue("$id", id);
-        command.Parameters.AddWithValue("$ram", ram);
-        command.Parameters.AddWithValue("$processor", processor);
-
-        command.ExecuteNonQuery();
-        connection.Close();
+       var computer = new Computer(id, ram, processor);
+       computerRepository.Save(computer);
     }
 }
