+++ Repositories/ComputerRepository.cs
@@ -15,6 +15,7 @@ class ComputerRepository
     }
 
     public List<Computer> GetAll()
+
     {
         var connection = new SqliteConnection("Data Source=database.db");
          connection.Open();
@@ -39,6 +40,20 @@ class ComputerRepository
          return computers;
     }
 
-//public Computer Save(Computer)
+public void Save(Computer computer)
+{
+    var connection = new SqliteConnection("Data Source=database.db");
+        connection.Open();
+
+        var command = connection.CreateCommand();
+        command.CommandText = "INSERT INTO Computers VALUES ($id, $ram, $processor);"
+        ;
+        command.Parameters.AddWithValue("$id", computer.Id);
+        command.Parameters.AddWithValue("$ram", computer.Ram);
+        command.Parameters.AddWithValue("$processor", computer.Processor);
+
+        command.ExecuteNonQuery();
+        connection.Close();
+}
   
 }
